Refactoring and bug fixes.

+++ oauth/access_token.go
@@ -29,7 +29,7 @@ func (s *Service) GrantAccessToken(client *Client, user *User, scope string) (*A
 // deleteExpiredAccessTokens deletes expired access tokens
 func (s *Service) deleteExpiredAccessTokens(client *Client, user *User) {
 	s.db.Unscoped().Where(AccessToken{
-		ClientID: util.IntOrNull(client.ID),
-		UserID:   util.IntOrNull(user.ID),
+		ClientID: util.IntOrNull(int64(client.ID)),
+		UserID:   util.IntOrNull(int64(user.ID)),
 	}).Where("expires_at <= ?", time.Now()).Delete(new(AccessToken))
 }
